+++ CHANGELOG.md
@@ -5,5 +5,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### changed
-- Adds an action to print a report (xlsx) from tree view of wizard "Inventory at Date" of stock_account module.
+### added
+- Adds an action to print a report (xlsx) of Inventory at date presenting the state of stock of products with variants adding a column by value of attribute. The attributes to present in report should be seted in the system parameter "inventory_at_date_report_xlsx_variants_attrs". This report is available in the contextual menu from tree view of wizard "Inventory at Date" of stock_account module.